+++ packages/contracts/src/framework/plugin/setup/PluginSetup.sol
@@ -3,40 +3,37 @@
 pragma solidity ^0.8.8;
 
 import {ERC165} from "@openzeppelin/contracts/utils/introspection/ERC165.sol";
-import {ERC165Checker} from "@openzeppelin/contracts/utils/introspection/ERC165Checker.sol";
 import {Clones} from "@openzeppelin/contracts/proxy/Clones.sol";
 
 import {IProtocolVersion, ProtocolVersion} from "../../../utils/protocol/ProtocolVersion.sol";
 import {PermissionLib} from "../../../core/permission/PermissionLib.sol";
-import {createERC1967Proxy as createERC1967} from "../../../utils/Proxy.sol";
 import {IPluginSetup} from "./IPluginSetup.sol";
 
 /// @title PluginSetup
 /// @author Aragon Association - 2022-2023
 /// @notice An abstract contract that developers have to inherit from to write the setup of a plugin.
 /// @custom:security-contact sirt@aragon.org
-abstract contract PluginSetup is ERC165, IPluginSetup, ProtocolVersion {
+abstract contract PluginSetup is ERC165, IPluginSetup {
+    using Clones for address;
+
+    /// @notice Thrown when trying to update a non-upgradeable plugin.
+    error UpdatesUnsupported();
+
+    /// @notice The address of plugin implementation contract to create [ERC-1167](https://eips.ethereum.org/EIPS/eip-1167) clones from.
+    address public immutable implementation;
+
+    /// @notice The contract constructor, that deploys the `Multisig` plugin logic contract.
+    constructor(address _implementation) {
+        implementation = _implementation;
+    }
+
     /// @inheritdoc IPluginSetup
     function prepareUpdate(
-        address _dao,
-        uint16 _currentBuild,
-        SetupPayload calldata _payload
-    )
-        external
-        virtual
-        override
-        returns (bytes memory initData, PreparedSetupData memory preparedSetupData)
-    {}
-
-    /// @notice A convenience function to create an [ERC-1967](https://eips.ethereum.org/EIPS/eip-1967) proxy contract pointing to an implementation and being associated to a DAO.
-    /// @param _implementation The address of the implementation contract to which the proxy is pointing to.
-    /// @param _data The data to initialize the storage of the proxy contract.
-    /// @return The address of the created proxy contract.
-    function createERC1967Proxy(
-        address _implementation,
-        bytes memory _data
-    ) internal returns (address) {
-        return createERC1967(_implementation, _data);
+        address,
+        uint16,
+        SetupPayload calldata
+    ) external pure override returns (bytes memory, PreparedSetupData memory) {
+        revert UpdatesUnsupported();
     }
 
     /// @notice Checks if this or the parent contract supports an interface by its ID.