+++ packages/contracts/src/plugins/governance/majority-voting/addresslist/AddresslistVotingSetup.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.8;
 import {IDAO} from "../../../../core/dao/IDAO.sol";
 import {DAO} from "../../../../core/dao/DAO.sol";
 import {PermissionLib} from "../../../../core/permission/PermissionLib.sol";
-import {PluginSetup, IPluginSetup} from "../../../../framework/plugin/setup/PluginSetup.sol";
+import {PluginSetupUpgradeable, IPluginSetup} from "../../../../framework/plugin/setup/PluginSetupUpgradeable.sol";
 import {MajorityVotingBase} from "../MajorityVotingBase.sol";
 import {AddresslistVoting} from "./AddresslistVoting.sol";
 
@@ -13,14 +13,9 @@ import {AddresslistVoting} from "./AddresslistVoting.sol";
 /// @author Aragon Association - 2022-2023
 /// @notice The setup contract of the `AddresslistVoting` plugin.
 /// @custom:security-contact sirt@aragon.org
-contract AddresslistVotingSetup is PluginSetup {
-    /// @notice The address of `AddresslistVoting` plugin logic contract to be used in creating proxy contracts.
-    AddresslistVoting private immutable addresslistVotingBase;
-
+contract AddresslistVotingSetup is PluginSetupUpgradeable {
     /// @notice The contract constructor, that deploys the `AddresslistVoting` plugin logic contract.
-    constructor() {
-        addresslistVotingBase = new AddresslistVoting();
-    }
+    constructor() PluginSetupUpgradeable(address(new AddresslistVoting())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
@@ -33,9 +28,15 @@ contract AddresslistVotingSetup is PluginSetup {
 
         // Prepare and Deploy the plugin proxy.
         plugin = createERC1967Proxy(
-            address(addresslistVotingBase),
-            abi.encodeCall(AddresslistVoting.initialize, (IDAO(_dao), votingSettings, members))
+            implementation,
+            abi.encodeWithSelector(
+                AddresslistVoting.initialize.selector,
+                _dao,
+                votingSettings,
+                members
+            )
         );
+        AddresslistVoting addresslistVoting = AddresslistVoting(implementation);
 
         // Prepare permissions
         PermissionLib.MultiTargetPermission[]
@@ -48,7 +49,7 @@ contract AddresslistVotingSetup is PluginSetup {
             where: plugin,
             who: _dao,
             condition: PermissionLib.NO_CONDITION,
-            permissionId: addresslistVotingBase.UPDATE_ADDRESSES_PERMISSION_ID()
+            permissionId: addresslistVoting.UPDATE_ADDRESSES_PERMISSION_ID()
         });
 
         permissions[1] = PermissionLib.MultiTargetPermission({
@@ -56,7 +57,7 @@ contract AddresslistVotingSetup is PluginSetup {
             where: plugin,
             who: _dao,
             condition: PermissionLib.NO_CONDITION,
-            permissionId: addresslistVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
+            permissionId: addresslistVoting.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         });
 
         permissions[2] = PermissionLib.MultiTargetPermission({
@@ -64,7 +65,7 @@ contract AddresslistVotingSetup is PluginSetup {
             where: plugin,
             who: _dao,
             condition: PermissionLib.NO_CONDITION,
-            permissionId: addresslistVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
+            permissionId: addresslistVoting.UPGRADE_PLUGIN_PERMISSION_ID()
         });
 
         // Grant `EXECUTE_PERMISSION` of the DAO to the plugin.
@@ -84,6 +85,8 @@ contract AddresslistVotingSetup is PluginSetup {
         address _dao,
         SetupPayload calldata _payload
     ) external view returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+        AddresslistVoting addresslistVoting = AddresslistVoting(_payload.plugin);
+
         // Prepare permissions
         permissions = new PermissionLib.MultiTargetPermission[](4);
 
@@ -93,7 +96,7 @@ contract AddresslistVotingSetup is PluginSetup {
             where: _payload.plugin,
             who: _dao,
             condition: PermissionLib.NO_CONDITION,
-            permissionId: addresslistVotingBase.UPDATE_ADDRESSES_PERMISSION_ID()
+            permissionId: addresslistVoting.UPDATE_ADDRESSES_PERMISSION_ID()
         });
 
         permissions[1] = PermissionLib.MultiTargetPermission({
@@ -101,7 +104,7 @@ contract AddresslistVotingSetup is PluginSetup {
             where: _payload.plugin,
             who: _dao,
             condition: PermissionLib.NO_CONDITION,
-            permissionId: addresslistVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
+            permissionId: addresslistVoting.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         });
 
         permissions[2] = PermissionLib.MultiTargetPermission({
@@ -109,7 +112,7 @@ contract AddresslistVotingSetup is PluginSetup {
             where: _payload.plugin,
             who: _dao,
             condition: PermissionLib.NO_CONDITION,
-            permissionId: addresslistVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
+            permissionId: addresslistVoting.UPGRADE_PLUGIN_PERMISSION_ID()
         });
 
         permissions[3] = PermissionLib.MultiTargetPermission({
@@ -119,10 +122,5 @@ contract AddresslistVotingSetup is PluginSetup {
             condition: PermissionLib.NO_CONDITION,
             permissionId: DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         });
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view returns (address) {
-        return address(addresslistVotingBase);
     }
 }