+++ packages/contracts/src/test/plugin/UUPSUpgradeable/PluginUUPSUpgradeableSetupMock.sol
@@ -4,24 +4,20 @@ pragma solidity ^0.8.8;
 
 import {PermissionLib} from "../../../core/permission/PermissionLib.sol";
 import {IDAO} from "../../../core/dao/IDAO.sol";
-import {PluginSetup} from "../../../framework/plugin/setup/PluginSetup.sol";
+import {PluginSetupUpgradeable} from "../../../framework/plugin/setup/PluginSetupUpgradeable.sol";
 import {IPluginSetup} from "../../../framework/plugin/setup/IPluginSetup.sol";
 import {mockPermissions, mockHelpers, mockPluginProxy} from "../PluginMockData.sol";
 import {PluginUUPSUpgradeableV1Mock, PluginUUPSUpgradeableV2Mock, PluginUUPSUpgradeableV3Mock} from "./PluginUUPSUpgradeableMock.sol";
 
-contract PluginUUPSUpgradeableSetupV1Mock is PluginSetup {
-    address internal pluginBase;
-
-    constructor() {
-        pluginBase = address(new PluginUUPSUpgradeableV1Mock());
-    }
+contract PluginUUPSUpgradeableSetupV1Mock is PluginSetupUpgradeable {
+    constructor() PluginSetupUpgradeable(address(new PluginUUPSUpgradeableV1Mock())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
     ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
-        plugin = mockPluginProxy(pluginBase, _dao);
+        plugin = mockPluginProxy(implementation, _dao);
         preparedSetupData.helpers = mockHelpers(2);
         preparedSetupData.permissions = mockPermissions(0, 2, PermissionLib.Operation.Grant);
     }
@@ -33,11 +29,6 @@ contract PluginUUPSUpgradeableSetupV1Mock is PluginSetup {
     ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
         (_dao, _payload);
         permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
-    }
-
-    /// @inheritdoc IPluginSetup
-    function implementation() external view virtual override returns (address) {
-        return address(pluginBase);
     }
 }
 
@@ -53,17 +44,15 @@ contract PluginUUPSUpgradeableSetupV1MockBad is PluginUUPSUpgradeableSetupV1Mock
     }
 }
 
-contract PluginUUPSUpgradeableSetupV2Mock is PluginUUPSUpgradeableSetupV1Mock {
-    constructor() {
-        pluginBase = address(new PluginUUPSUpgradeableV2Mock());
-    }
+contract PluginUUPSUpgradeableSetupV2Mock is PluginSetupUpgradeable {
+    constructor() PluginSetupUpgradeable(address(new PluginUUPSUpgradeableV2Mock())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
     ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
-        plugin = mockPluginProxy(pluginBase, _dao);
+        plugin = mockPluginProxy(implementation, _dao);
         preparedSetupData.helpers = mockHelpers(2);
         preparedSetupData.permissions = mockPermissions(0, 2, PermissionLib.Operation.Grant);
     }
@@ -87,19 +76,26 @@ contract PluginUUPSUpgradeableSetupV2Mock is PluginUUPSUpgradeableSetupV1Mock {
             preparedSetupData.permissions = mockPermissions(1, 2, PermissionLib.Operation.Grant);
         }
     }
+
+    /// @inheritdoc IPluginSetup
+    function prepareUninstallation(
+        address _dao,
+        SetupPayload calldata _payload
+    ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+        (_dao, _payload);
+        permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
+    }
 }
 
-contract PluginUUPSUpgradeableSetupV3Mock is PluginUUPSUpgradeableSetupV2Mock {
-    constructor() {
-        pluginBase = address(new PluginUUPSUpgradeableV3Mock());
-    }
+contract PluginUUPSUpgradeableSetupV3Mock is PluginSetupUpgradeable {
+    constructor() PluginSetupUpgradeable(address(new PluginUUPSUpgradeableV3Mock())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
     ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
-        plugin = mockPluginProxy(pluginBase, _dao);
+        plugin = mockPluginProxy(implementation, _dao);
         preparedSetupData.helpers = mockHelpers(3);
         preparedSetupData.permissions = mockPermissions(0, 3, PermissionLib.Operation.Grant);
     }
@@ -130,6 +126,15 @@ contract PluginUUPSUpgradeableSetupV3Mock is PluginUUPSUpgradeableSetupV2Mock {
             preparedSetupData.permissions = mockPermissions(2, 3, PermissionLib.Operation.Grant);
         }
     }
+
+    /// @inheritdoc IPluginSetup
+    function prepareUninstallation(
+        address _dao,
+        SetupPayload calldata _payload
+    ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+        (_dao, _payload);
+        permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
+    }
 }
 
 /// @dev With this plugin setup, the plugin base implementation doesn't change.
@@ -137,17 +142,15 @@ contract PluginUUPSUpgradeableSetupV3Mock is PluginUUPSUpgradeableSetupV2Mock {
 /// which uses the same base implementation(doesn't update the logic contract)
 /// but applies new/modifier permissions on it.
 
-contract PluginUUPSUpgradeableSetupV4Mock is PluginUUPSUpgradeableSetupV3Mock {
-    constructor(address _pluginUUPSUpgradeableV3) {
-        pluginBase = _pluginUUPSUpgradeableV3;
-    }
+contract PluginUUPSUpgradeableSetupV4Mock is PluginSetupUpgradeable {
+    constructor() PluginSetupUpgradeable(address(new PluginUUPSUpgradeableV3Mock())) {}
 
     /// @inheritdoc IPluginSetup
     function prepareInstallation(
         address _dao,
         bytes memory
     ) public virtual override returns (address plugin, PreparedSetupData memory preparedSetupData) {
-        plugin = mockPluginProxy(pluginBase, _dao);
+        plugin = mockPluginProxy(implementation, _dao);
         preparedSetupData.helpers = mockHelpers(3);
         preparedSetupData.permissions = mockPermissions(0, 3, PermissionLib.Operation.Grant);
     }
@@ -175,9 +178,44 @@ contract PluginUUPSUpgradeableSetupV4Mock is PluginUUPSUpgradeableSetupV3Mock {
         // returning the `initData` that will be called upon `upradeToAndCall` by plugin setup processor.
         // NOTE that dev is free to do what he wishes.
         else if (_currentBuild == 1 || _currentBuild == 2) {
-            (initData, preparedSetupData) = super.prepareUpdate(_dao, _currentBuild, _payload);
+            (initData, preparedSetupData) = prepareUpdateAlt(_dao, _currentBuild, _payload);
             // Even for this case, dev might decide to modify the permissions..
             preparedSetupData.permissions = mockPermissions(4, 5, PermissionLib.Operation.Grant);
         }
+    }
+
+    function prepareUpdateAlt(
+        address _dao,
+        uint16 _currentBuild,
+        SetupPayload calldata _payload
+    ) internal pure returns (bytes memory initData, PreparedSetupData memory preparedSetupData) {
+        (_dao, _payload);
+
+        // Update from V1
+        if (_currentBuild == 1) {
+            preparedSetupData.helpers = mockHelpers(3);
+            initData = abi.encodeWithSelector(
+                PluginUUPSUpgradeableV3Mock.initializeV1toV3.selector
+            );
+            preparedSetupData.permissions = mockPermissions(1, 3, PermissionLib.Operation.Grant);
+        }
+
+        // Update from V2
+        if (_currentBuild == 2) {
+            preparedSetupData.helpers = mockHelpers(3);
+            initData = abi.encodeWithSelector(
+                PluginUUPSUpgradeableV3Mock.initializeV2toV3.selector
+            );
+            preparedSetupData.permissions = mockPermissions(2, 3, PermissionLib.Operation.Grant);
+        }
+    }
+
+    /// @inheritdoc IPluginSetup
+    function prepareUninstallation(
+        address _dao,
+        SetupPayload calldata _payload
+    ) external virtual override returns (PermissionLib.MultiTargetPermission[] memory permissions) {
+        (_dao, _payload);
+        permissions = mockPermissions(0, 1, PermissionLib.Operation.Revoke);
     }
 }